Changed entites & errors & executor to micro-lib
