+++ backend/libs/errors/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "executor"
+name = "errors"
 version = "0.1.0"
 authors = ["Hiroaki Yamamoto <hiroaki@hysoftware.net>"]
 edition = "2018"
@@ -7,10 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-entities = { path = "../entities" }
-errors = { path = "../../errors" }
-types = { path = "../../types" }
-
-async-trait = "0.1.42"
-bson = "1.2.0"
-futures-core = "0.3.12"
+serde = {version = "1.0.123", features=["derive"]}
+slog = "2.7.0"
+slog_derive = "0.2.0"
+url = "2.2.0"
