+++ backend/libs/exchanges/src/lib.rs
@@ -1,14 +1,14 @@
 pub mod binance;
 mod casting;
-mod entities;
 mod keychain;
 mod traits;
 mod types;
 
 pub mod errors;
 
-pub use crate::entities::{APIKey, KlineCtrl, ListSymbolStream, OrderOption};
-pub use crate::keychain::KeyChain;
-pub use crate::traits::{
-  HistoryFetcher, HistoryRecorder, SymbolFetcher, TradeObserver, UserStream,
-};
+// pub use crate::entities::{APIKey, KlineCtrl, ListSymbolStream, OrderOption};
+// pub use crate::keychain::KeyChain;
+// pub use crate::traits::{
+//   HistoryFetcher, HistoryRecorder, SymbolFetcher, TradeObserver, UserStream,
+// };
+pub use ::entities::*;
