+++ backend/libs/rulesets/entities/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "executor"
+name = "entities"
 version = "0.1.0"
 authors = ["Hiroaki Yamamoto <hiroaki@hysoftware.net>"]
 edition = "2018"
@@ -7,10 +7,9 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-entities = { path = "../entities" }
-errors = { path = "../../errors" }
-types = { path = "../../types" }
+rpc = { path = "../../rpc" }
 
-async-trait = "0.1.42"
-bson = "1.2.0"
 futures-core = "0.3.12"
+num-traits = "0.2.14"
+serde = { version = "1.0.123", features = ["derive"] }
+bson = "1.2.0"
