+++ backend/libs/rulesets/entities/src/lib.rs
@@ -5,15 +5,15 @@ mod kline;
 mod order;
 mod order_option;
 
-pub(crate) use self::apikey::APIKeyEvent;
+pub use self::apikey::APIKeyEvent;
 pub use self::apikey::{APIKey, APIKeyInternal};
 pub use self::book_ticker::BookTicker;
 pub use self::execution::{ExecutionResult, ExecutionType};
 pub use self::kline::KlineCtrl;
-pub(crate) use self::order::{Order, OrderInner};
+pub use self::order::{Order, OrderInner};
 pub use self::order_option::OrderOption;
 
-use ::futures::stream::BoxStream;
+use ::futures_core::stream::BoxStream;
 use ::rpc::entities::SymbolInfo;
 
 pub type ListSymbolStream<'a> = BoxStream<'a, SymbolInfo>;
